requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pgvector==0.2.4

# Data processing
pandas==2.1.4
numpy==1.24.3

# OpenAI & Embeddings
openai==1.3.7
scikit-learn==1.3.2

# HTTP requests
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# CORS middleware
python-multipart==0.0.6